# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a server to support siloed inbox creation, and mailgun interactions for the [bip44Mail-client](https://github.com/bip44Mail/bip44Mail-client)",
    "version": "1.0.0",
    "title": "bip44Mail",
    "termsOfService": "",
    "contact": {
      "email": "creativemedicine@protonmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3000",
  "basePath": "/v2",
  "tags": [
    {
      "name": "sending",
      "description": "Everything about sending mail."
    },
    {
      "name": "receiving",
      "description": "Everything about receiving mail."
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/send": {
      "post": {
        "tags": [
          "sending"
        ],
        "summary": "Send an email",
        "description": "Sends an email if the provided signature matches the from address.",
        "operationId": "sendMail",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Email info and an ECDSA signature.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SendRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success."
          },
          "500": {
            "description": "Failure."
          }
        }
      }
    },
    "/inbox": {
      "post": {
        "tags": [
          "receiving"
        ],
        "summary": "Create a new inbox.",
        "description": "Creates a new inbox.",
        "operationId": "createInbox",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "ECDSA signature and address to map to an inbox.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateInboxRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success."
          },
          "500": {
            "description": "Failure."
          }
        }
      }
    }
  },
  "definitions": {
    "SendRequest": {
      "type": "object",
      "required": [
        "signature",
        "from",
        "to",
        "subject",
        "message"
      ],
      "properties": {
        "signature": {
          "type": "string"
        },
        "from": {
          "type": "string"
        },
        "to": {
          "type": "string"
        },
        "subject": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "CreateInboxRequest": {
      "type": "object",
      "required": [
        "signature",
        "address"
      ],
      "properties": {
        "signature": {
          "type": "string"
        },
        "address": {
          "type": "string"
        }
      }
    }
  }
}